Move update function to item card component, only update changed values
