+++ client/src/component/item/ItemCard.jsx
@@ -10,11 +10,11 @@ import TagList from "./../tag/TagList"
 import Pin from "./../pin/Pin"
 import { useUserAuth } from "./../context/UserAuthProvider"
 import { useCurrentInput } from "./../context/CurrentInputProvider"
-import { updateMetaImage } from "./../../logic/api/item"
+import { updateItem, updateMetaImage } from "./../../logic/api/item"
 
 import "./ItemCard.css"
 
-function ItemCard({ item, createItemModalRef, handleItemPinClick }) {
+function ItemCard({ item, createItemModalRef, triggerItemLoad }) {
     const prefersColorScheme = usePrefersColorScheme()
     const isDarkMode = prefersColorScheme === "dark"
 
@@ -43,6 +43,25 @@ function ItemCard({ item, createItemModalRef, handleItemPinClick }) {
 
         createItemModalRef.current.setEditId(itemId)
         createItemModalRef.current.open()
+    }
+    const handleItemPinClick = async (item) => {
+        try {
+            await updateItem(
+                item._id,
+                {
+                    isPinned: !item.isPinned,
+                },
+                workspace,
+                accessToken
+            )
+        } catch (ex) {
+            toast.error(ex.message)
+
+            return
+        }
+        toast.success(`Item ${!item.isPinned ? "pinned" : "unpinned"}!`)
+
+        triggerItemLoad()
     }
     const handleItemUpdateMetaImageClick = async (e, itemId) => {
         e.preventDefault()
